+++ crates/obscura-browser/src/lib.rs
@@ -5,3 +5,4 @@ pub mod lifecycle;
 pub use page::{Page, PageError};
 pub use context::BrowserContext;
 pub use lifecycle::{LifecycleState, WaitUntil};
+pub use obscura_js::HTML_TO_MARKDOWN_JS;
